06-text-generation-apps/python/requirements.txt
openai==1.55.3
python-dotenv==1.0.1
httpx==0.28.0
tiktoken==0.4.0
